requirements.txt
appdirs==1.4.4
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
packaging==23.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
